In the workouts controller, added the 'new_with_search' method to provide a search facility for the name of exercises.
In the edit method, added the @exs and @allexs arrays to hold the list of current exercises of a workout and all available exercises respectively.

While creating a workout, changed the code to capture the relationship with exercises and populating the attributes of the Exercise_workouts table.

For the practice_workout method, added the functionality for a student to take a workout by storing the remaining exercises for a workout in the session variable.

Permitting the other parameters necessary for workouts to come through in 'workout_params'.