Changed the very nature of exercise. Each exercise is now a version of its base exercise, which is generally represented by the variable 'basex'. So when an exercise is created through the forms, both the BaseExercise and Exercise object are created; the latter as the first version.
Now checking for the nullability of the pesky tags_attributes.

Scoring is now based on the points for each exercise as determined by its associated workout.

Added the mechanism through params[:wexes] (containing the remaining exercises in a workout to be attempted) to practice a workout by going through each workout.

Added a new method to record score for the exercises in a workout using the new WorkoutScore model.

For exercise practice, to facilitate the display of feedback, rendering of 'three_columns' is used.